refactor: move styles files to separate folder
