+++ src/styles/docs-styles.ts
