+++ src/styles/highlight-styles.ts
